requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=10.0.0

# Diffusion models and LoRA
diffusers>=0.25.0
transformers>=4.35.0
accelerate>=0.25.0
peft>=0.7.0

# Training and logging
tensorboard>=2.15.0
tqdm>=4.66.0
pyyaml>=6.0

# Data processing
opencv-python>=4.8.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Optional: for better performance
# ninja  # For faster compilation of CUDA extensions
# xformers  # For memory-efficient attention (if compatible)
